fix: migrate real_portfolio quantity/price from double to BigDecimal
jf_investments.quantity/purchase_price were double precision — real money
must never use binary floating point, since rounding error compounds
across lots and can drift a computed value onto the wrong side of an
achievement threshold (portfolio_10k/50k, positive_return). Converts the
whole ledger chain (Investment domain/entity, UserPositionCalculator,
InvestmentLotDTO, PortfolioSummaryDTO, AllocationSliceDTO,
PortfolioHistoryPointDTO, AchievementContext/Catalog) to BigDecimal, scale
2 HALF_UP for money, scale 6 for quantity, matching simulated_portfolio's
own convention.

Flyway V24 uses "SET DATA TYPE ... USING" for the column type change —
verified empirically against this project's real H2 2.4.240 that this
exact syntax is accepted by both H2 and PostgreSQL, unlike H2's no-USING
form or Postgres's bare TYPE keyword. No rows dropped, every existing
value cast in place.

Deliberately left as Double, a separate lower-stakes chain: Dividend Radar
(DividendDTO/DividendRadarEntryDTO) and AssetDetailsResponseDTO's ~40
external market-fundamentals display fields (never persisted, never used
in a balance-affecting calculation this backend owns).

Full suite 877/877 green.

+++ src/main/java/com/jf/PetApp/application/gamification/achievement/AchievementContext.java
@@ -1,5 +1,6 @@
 package com.jf.PetApp.application.gamification.achievement;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
@@ -8,17 +9,18 @@ import java.time.LocalDate;
  * already use ({@code GetPortfolioHoldingsUseCase},
  * {@code GetPortfolioSummaryUseCase}, {@code GetPortfolioAllocationUseCase}),
  * so an achievement can never be unlocked on anything the server hasn't
- * independently verified.
+ * independently verified. Money fields are {@link BigDecimal}, matching the
+ * real_portfolio calculation chain they're built from.
  */
 public record AchievementContext(
         boolean hasHoldings,
-        double currentValue,
-        double totalGain,
+        BigDecimal currentValue,
+        BigDecimal totalGain,
         int distinctTypeCount,
         int distinctFundsTickerCount,
         LocalDate firstPurchaseDate,
-        double monthlyPassiveIncomeEstimate,
-        double annualPassiveIncomeEstimate) {
+        BigDecimal monthlyPassiveIncomeEstimate,
+        BigDecimal annualPassiveIncomeEstimate) {
 
     public long daysSinceFirstPurchase(LocalDate today) {
         if (firstPurchaseDate == null) {
