+++ src/main/java/com/jf/PetApp/application/investment/dto/AllocationSliceDTO.java
@@ -2,9 +2,11 @@ package com.jf.PetApp.application.investment.dto;
 
 import com.jf.PetApp.core.domain.enums.InvestmentType;
 
+import java.math.BigDecimal;
+
 public record AllocationSliceDTO(
     InvestmentType type,
-    Double currentValue,
-    Double portfolioPercent
+    BigDecimal currentValue,
+    BigDecimal portfolioPercent
 ) {
 }
