+++ src/test/java/com/jf/PetApp/application/investment/dto/PortfolioSummaryDTOTest.java
@@ -2,18 +2,21 @@ package com.jf.PetApp.application.investment.dto;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class PortfolioSummaryDTOTest {
 
     @Test
     void accessorsReturnConstructedValues() {
-        PortfolioSummaryDTO dto = new PortfolioSummaryDTO(1000.0, 1200.0, 200.0, 20.0, 5);
+        PortfolioSummaryDTO dto = new PortfolioSummaryDTO(
+                BigDecimal.valueOf(1000.0), BigDecimal.valueOf(1200.0), BigDecimal.valueOf(200.0), BigDecimal.valueOf(20.0), 5);
 
-        assertEquals(1000.0, dto.investedCapital());
-        assertEquals(1200.0, dto.currentValue());
-        assertEquals(200.0, dto.totalGain());
-        assertEquals(20.0, dto.totalGainPercent());
+        assertEquals(BigDecimal.valueOf(1000.0), dto.investedCapital());
+        assertEquals(BigDecimal.valueOf(1200.0), dto.currentValue());
+        assertEquals(BigDecimal.valueOf(200.0), dto.totalGain());
+        assertEquals(BigDecimal.valueOf(20.0), dto.totalGainPercent());
         assertEquals(5, dto.totalAssets());
     }
 }
